de-bounce the changecount by allowing batching of multiple changes
